Rename C++ headers to .hpp and make checkpatch only check C code.
