+++ src/grpc/dp_grpc_conv.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and IronCore contributors
 // SPDX-License-Identifier: Apache-2.0
 
-#include "grpc/dp_grpc_conv.h"
+#include "grpc/dp_grpc_conv.hpp"
 #include "dp_error.h"
 #include "dp_log.h"
