+++ src/grpc/dp_grpc_health.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2024 SAP SE or an SAP affiliate company and IronCore contributors
 // SPDX-License-Identifier: Apache-2.0
 
-#include "grpc/dp_grpc_health.h"
+#include "grpc/dp_grpc_health.hpp"
 #include "../proto/dpdk.grpc.pb.h"
 
 void HealthService::SetServingStatus(grpc::health::v1::HealthCheckResponse::ServingStatus newStatus)
